Version v2.50.61-beta7 released!
-Adding 64-bit Integers support for possible AMD64
- Enums modernised
- Replaced some C casts with C++ casts
- Now supporting C++17 but still using C99
- Code optimised to reduce possible mem leaks and modernised loops
- Homesite now as "adminmod.de"

+++ S-line/dlls/adminmod/amlibc.c
@@ -1,3 +1,5 @@
+// This is an open source non-commercial project. Dear PVS-Studio, please check it.
+// PVS-Studio Static Code Analyzer for C, C++, C#, and Java: http://www.viva64.com
 /*
  * ===========================================================================
  *
@@ -111,7 +113,7 @@ char *am_strcasestr(const char *haystack, const char *needle)
   if ( nl > hl ) return 0;
 
 	//for is only compatible with GCC C99 mode, use -std=c99 [APG]RoboCop[CL]
-  for ( int i = hl - nl + 1; i; --i ) {
+  for ( size_t i = hl - nl + 1; i; --i ) {
     if ( am_tolower(*haystack) == am_tolower(*needle) && !am_strcasecmp(haystack,needle)) return (char*)haystack;
 
     ++haystack;
